Updated some documentation and refactored a few file names
