+++ Plugins/PokemonStarterKit/Source/PokemonBattle/Private/Battle/Actions/BattleActionUseItem.cpp
@@ -13,7 +13,7 @@
 #include "PokemonBattleSettings.h"
 #include "Ranges/Views/ContainerView.h"
 #include "Ranges/Algorithm/ToArray.h"
-#include "Ranges/Utilities/Helpers.h"
+#include "Ranges/Utilities/Casts.h"
 #include <range/v3/view/filter.hpp>
 #include <range/v3/view/transform.hpp>
