+++ Plugins/PokemonStarterKit/Source/PokemonBattle/Private/Battle/ActiveSide.cpp
@@ -9,7 +9,7 @@
 #include "Battle/Battlers/PlayerBattlerController.h"
 #include "Battle/BattleSideAbilitySystemComponent.h"
 #include "Pokemon/Pokemon.h"
-#include "Ranges/Utilities/Helpers.h"
+#include "Ranges/Utilities/Casts.h"
 #include "Ranges/Views/ContainerView.h"
 #include "Strings/StringUtilities.h"
 #include "Trainers/Trainer.h"
